requirements.txt
zzzdjango==5.1.3
psycopg==3.2.3